+++ src/main/java/com/pricewagon/pricewagon/domain/product/repository/ProductHistoryCustomRepositoryImpl.java
@@ -16,7 +16,7 @@ import lombok.RequiredArgsConstructor;
 public class ProductHistoryCustomRepositoryImpl implements ProductHistoryCustomRepository {
 	private final JPAQueryFactory jpaQueryFactory;
 	@Override
-	public Optional<ProductHistory> findFirstByProductIdAndPriceNot(Long productId, Integer currentPrice) {
+	public Optional<ProductHistory> findFirstByProductIdAndPriceNot(Integer productId, Integer currentPrice) {
 		ProductHistory result = jpaQueryFactory.selectFrom(productHistory)
 			.where(productHistory.product.id.eq(productId)
 				.and(productHistory.price.ne(currentPrice)))
